requirements.txt
# Standard library modules used (no pip install needed): smtplib, email
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-api-python-client==2.126.0
gunicorn==21.2.0
blinker==1.9.0
click==8.3.2
et_xmlfile==2.0.0
Flask==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
nulltype==2.3.1
openpyxl==3.1.5
plaid-python==39.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
six==1.17.0
urllib3==2.6.3
Werkzeug==3.1.8
